predict.py
import os
import json
import io
from PIL import Image
import requests

class Predictor:
    def __init__(self):
        required_models = [
            "4x_NMKD-Siax_200k.pth",
            "FLUX.1-Turbo-Alpha 8steps .safetensors",
            "Flux1-Dev-SRPO-v1-Q8_0.gguf",
            "GFPGANv1.4.pth",
            "ae.safetensors",
            "clip_l.safetensors",
            "flux1-dev-Q8_0.gguf",
            "seedvr2_ema_7b_fp16.safetensors",
            "t5xxl_fp8_e4m3fn.safetensors"
        ]
        self.model_errors = [m for m in required_models if not os.path.exists(m)]
        if self.model_errors:
            raise RuntimeError(f"Faltan modelos esenciales: {self.model_errors}")
        try:
            with open("workflow_api.json") as f:
                self.workflow = json.load(f)
        except Exception as e:
            raise RuntimeError(f"No se puede cargar workflow_api.json: {e}")

    def load_image(self, image=None, image_url=None):
        if image:
            return Image.open(image)
        if image_url:
            resp = requests.get(image_url, stream=True)
            resp.raise_for_status()
            return Image.open(io.BytesIO(resp.content))
        raise RuntimeError("Debes proporcionar image o image_url.")

    def predict(self, image=None, image_url=None, **kwargs):
        try:
            img = self.load_image(image, image_url)
            img = img.convert("RGB")
            # Aquí deberías llamar el workflow real,
            # para ejemplo solo guardamos input como output
            output_path = "output.jpg"
            img.save(output_path)
            return output_path   # Devuelve la ruta al archivo generado (Replicate lo sirve como output)
        except Exception as e:
            return {"status": "ERROR", "error": str(e)}
